Create new classes for refactoring
