+++ build.gradle
@@ -45,7 +45,6 @@ dependencies {
 
     // SQLite for database
     compile group:'org.xerial', name:'sqlite-jdbc', version:'3.30.1'
-    driver 'org.xerial:sqlite-jdbc:3.30.1'
 
     // Mockito for testing
     testCompile "org.mockito:mockito-core:1.+"
@@ -55,8 +54,3 @@ application {
     // Define the main class for the application.
     mainClassName = 'planner.Main'
 }
-
-task createDB {
-    println "foo"
-}
-installDist.dependsOn createDB
