+++ src/main/java/planner/Database.java
@@ -1,0 +1,7 @@
+package planner;
+
+/**
+ * Handles the DB communications.
+ */
+public class Database {
+}
