+++ src/main/java/planner/entities/Attack.java
@@ -1,4 +1,4 @@
-package planner;
+package planner.entities;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
