+++ src/main/java/planner/entities/Operation.java
@@ -1,0 +1,7 @@
+package planner.entities;
+
+/**
+ * This class represents a whole planned operation in memory and offers load and save methods for long-term storage.
+ */
+public class Operation {
+}
