+++ src/main/resources/planner/entities/images/delete.gif
