+++ src/main/resources/planner/entities/images/teutons.gif
